requirements.txt
# numeric
numpy>=1.19.5
scipy>=1.7.1

# plotting
matplotlib>=3.3.4

# dataframes
pandas>=1.3.2

# torch for cpu
# see https://pytorch.org/ for CUDA
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu
torchvision==0.10.1+cpu
torchaudio==0.9.1

# logging
mlflow==1.23.1

# more ML
scikit-learn==0.24.2
pytorch-lightning==1.4.7

# config
hydra-core>=1.1.2

# others
requests>=2.26.0
tqdm>=4.62.2
netron>=5.3.3